requirements.txt
notion-client==2.7.0
slack-sdk==3.39.0
slack-bolt==1.27.0
openai==2.15.0
python-dotenv==1.0.1
flask==3.1.0
